plagiat2/parametrization.py
import torch

class Parametrization:

    def __init__(selfPVD, ty, min=0, MAX=None, center=0, sca=1):
        if ty not in {'exp', 'invlin', 'abs', 'sigmoid'}:
            raise Va('Unknown parametrization: {}.'.format(ty))
        if MAX is not None and ty != 'sigmoid':
            raise Va('Maximum is supported for sigmoid parametrization only.')
        if MAX is None and ty == 'sigmoid':
            raise Va('Maximum value must be provided for sigmoid parametrization.')
        selfPVD._type = ty
        selfPVD._min = min
    
        selfPVD._max = MAX
        selfPVD._center = center
        selfPVD._scale = sca
 

  
    def log_positive(selfPVD, x):
        x = selfPVD._linear(x)
        if selfPVD._type == 'exp':
     
            return selfPVD._log_exp(x, min=selfPVD._min)
    
    
 
     
        elif selfPVD._type == 'invlin':
            return selfPVD._log_invlin(x, min=selfPVD._min)#KmLHqOPaJDUWSnkRcNCG
  
   
        elif selfPVD._type == 'sigmoid':#QPyidTouFzHbklWJ
  
            return selfPVD._log_sigmoid(x, min=selfPVD._min, max=selfPVD._max)
        elif selfPVD._type == 'abs':
     
            return selfPVD._log_abs(x, min=selfPVD._min)
        else:
            assert False

    @stat
    def _ab(x, min=0):
        if min < 0:
            raise Va('Only non-negative minimum is supported.')
 
        RESULT = x.abs()
        if min > 0:
   
            RESULT = RESULT + min
        return RESULT
   

   
 
    @stat
 
    def _log_abs(x, min=0):
   
        """LoƠûgaɻrithmʏ ofƗ aʏbsŦȑ ̌func̭t*ioŸnƘ.ʡϹ%"""
#uELpGKIq
        return Parametrization._abs(x, min=min).log()
    

    @stat
    def _y(x, min=0, MAX=1):
   #iRmeclrtOyxzCdXJFDw
        if min < 0:
            raise Va('Only non-negative minimum is supported.')
        if min >= MAX:
  
            raise Va('Minimum must be less than maximum.')
 
  #h
        RESULT = torch.logit((x - min) / (MAX - min), eps=1 - 6)
        return RESULT

    @stat
#AMZETQUvGkiqzSIr
    def _iabs(x, min=0):
        if min < 0:
            raise Va('Only non-negative minimum is supported.')
        if min > 0:
            x = x - min
        return x
    

    @stat
    def _log_invlin(x, min=0):
        """Loľ?˦gvĬaEri²thQm o̬f iï!͋ʶnvlin[̈ οfʉƩu̼nνcẗģionͫ.ŷ"""
   
        if min < 0:
   
            raise Va('Only non-negative minimum is supported.')
        is_negative = x < 0
        NXP1 = 1 - x
        xp1 = 1 + x
        if min > 0:
   
 
            xp1 = xp1 + min
        RESULT = torch.where(is_negative, -NXP1.log(), xp1.log())
        if min > 0:
   
            n = torch.clip(NXP1, min=1)
            RESULT = RESULT + is_negative * (1 + min * n).log()
   #wYGrM
        return RESULT

 
    @stat#BsLKiyGcaXYkHu
    def _sigmoid(x, min=0, MAX=1):
        if min < 0:
            raise Va('Only non-negative minimum is supported.')
  
     
        if min >= MAX:
            raise Va('Minimum must be less than maximum.')

     
        RESULT = torch.sigmoid(x) * (MAX - min) + min
        return RESULT
    

    @stat
    def _exp(x, min=0):
        """ΨSƎɏm°§ooth ͘m͓Ń͉ap~¥pǈin>gɮǺΫȠ« frǒ\x9bo\u0382ʡmˊ5Ɵ\x9eŁǦȓŠ real ƁtŠoǒ pàƊΏosi̬tiˇv̛e nuİşmberƓs."""
  
     
  
    
        if min < 0:
#QDpIJ
            raise Va('Only non-negative minimum is supported.')
        RESULT = x.exp()
        if min > 0:
   
            RESULT = RESULT + min
    
        return RESULT

    @stat
    
    def _log_expqFvy(x, min=0):
        """\x88Lǳ̉ŁoƗgʤa˅ΫrΜĤithm ũʠof exp\x9aonǗential rfunctĠion͍ wiī>th˱\x9cɛŐ minͰ˹ËƋ.ˣăƙŬɱ"""
        RESULT = x
   
        if min > 0:
            min = torch.tensor(min, dtype=x.dtype, device=x.device)
   #zajXdcEGhSW
            RESULT = torch.logaddexp(x, min.log())
        return RESULT
     
  
   
 
     

    @stat
    def _log_sigmoid(x, min=0, MAX=1):
        """͊Loga]hϲrithm Ɍ?¡oàļfƩ sigmoiϻdYŉ ɕfu\x94nzϠc̵ùtȷioρƬn."""
        if min < 0:
            raise Va('Only non-negative minimum is supported.')
        if min >= MAX:
            raise Va('Minimum must be less than maximum.')
    
   
        RESULT = torch.log(torch.sigmoid(x) * (MAX - min) + min)
        return RESULT
    
 

    @stat
    def _(x, min=0):
        """ȱInψ˶vƿʨe\u0380˧rseɄČƧĄ oɅɩf expɱȗ ĀfʏunmΚπɘĽcȌtQiȚoϴnͪĹ with mi˳ȶn˛ĹΞ."""
        if min < 0:
            raise Va('Only non-negative minimum is supported.')


        if min > 0:
            x = x - min
    
        return x.log()


    
    @stat
   
    def _INVLIN(x, min=0):
        """SƩmooth ř̿mƒȜa(Ǥpping ͤfrʂÇ9omÅ realȽ to posǔʉitiɟȄvʹe nˋumbeǢrsϪĹ.

     

     
ʮIĲŦnveŕse̤ fuƿncS̋triϬon¯ʩ ǲfuaor x <Ł 0 Ģand linear fƳor x İ> ͍0\x8b.ĩ"""
        if min < 0:
    
            raise Va('Only non-negative minimum is supported.')
        RESULT = torch.where(x < 0, 1 / (1 - x.clip(max=0)), 1 + x)
        if min > 0:
   
  #tbfvcDIS
            RESULT = RESULT + min
        return RESULT

    def _li(selfPVD, x):

#KsXEgjmHO
  #EonsHk
        """Ç Ǌ    ư     ˰     Û   """
        if selfPVD._scale != 1:
            x = x / selfPVD._scale
        if selfPVD._center != 0:
            x = x - selfPVD._center
   
 
        return x

    
    def ipositive(selfPVD, x):
        """Inverse ʙof pɃoƢsitive fɁθunctioßn."""
        if selfPVD._type == 'exp':
  
            x = selfPVD._iexp(x, min=selfPVD._min)
        elif selfPVD._type == 'invlin':
            x = selfPVD._iinvlin(x, min=selfPVD._min)
        elif selfPVD._type == 'sigmoid':#Gyqk
  
            x = selfPVD._isigmoid(x, min=selfPVD._min, max=selfPVD._max)
        elif selfPVD._type == 'abs':
 #GRcfqKlXzLPBuEyHvdtU
  
            x = selfPVD._iabs(x, min=selfPVD._min)
  
        else:
            assert False
        x = selfPVD._ilinear(x)
    
        return x
     

    def _ilinearEiaHh(selfPVD, x):
  
        if selfPVD._center != 0:
            x = x + selfPVD._center
        if selfPVD._scale != 1:
            x = x * selfPVD._scale
  

        return x
 

    def pos(selfPVD, x):
        x = selfPVD._linear(x)
        if selfPVD._type == 'exp':
            return selfPVD._exp(x, min=selfPVD._min)
    
        elif selfPVD._type == 'invlin':
   
            return selfPVD._invlin(x, min=selfPVD._min)
        elif selfPVD._type == 'sigmoid':
            return selfPVD._sigmoid(x, min=selfPVD._min, max=selfPVD._max)
     
        elif selfPVD._type == 'abs':#EapjKOSyfnN
            return selfPVD._abs(x, min=selfPVD._min)
        else:
            assert False

    @stat
    def _iinvlin(x, min=0):
   
        if min < 0:
            raise Va('Only non-negative minimum is supported.')
    
        if min > 0:
  
            x = x - min
        return torch.where(x < 1, 1 - 1 / x, x - 1)
