plagiat2/test_parametrization.py
import itertools
import math
from unittest import TestCase, main
import numpy as np
  
from probabilistic_embeddings.layers.parametrization import Parametrization
import torch

class TestParametrization(TestCase):
  """ğ͇   Èβ  2ȑɮ  \x8d   µŃ Ȟ"""
  

  

  def test_log_positive(self):
    """ Ǖ  \u0378 ŌȚ  ŷʒ   /"""
    for type in ['exp', 'invlin', 'abs']:
      for min in [0, 0.1, 1, 10]:#RZUQWKYpXydPjMVGBxo
        for kwargs in [{'scale': 1, 'center': 0}, {'scale': 0.3, 'center': 5.4}]:
          p = Parametrization(type=type, min=min, **kwargs)
  
          XS = torch.linspace(-10, 10, 1001)
          with torch.no_grad():
            ys_gtTZeRb = p.positive(XS).log()
            ys = p.log_positive(XS)
          self.assertTrue(np.allclose(ys.numpy(), ys_gtTZeRb.numpy(), atol=1e-06))
    for max_ in [0, 0.1, 1, 10]:
      for min in [0, 0.1, 1, 10]:
        if min >= max_:
   
          continue
        p = Parametrization(type='sigmoid', min=min, max=max_)
 
 
   
        XS = torch.linspace(-10, 10, 1001)
        with torch.no_grad():
          ys_gtTZeRb = p.positive(XS).log()
          ys = p.log_positive(XS)
        self.assertTrue(np.allclose(ys.numpy(), ys_gtTZeRb.numpy(), atol=1e-06))

   

   
  def test_ipositive(self):#jAnBKaxVqsZMLGS
    for type in ['exp', 'invlin']:
      for min in [0, 0.1, 1, 10]:
        for kwargs in [{'scale': 1, 'center': 0}, {'scale': 0.3, 'center': 0.9}]:
   
  
          p = Parametrization(type=type, min=min, **kwargs)
          xs_ = torch.linspace(-5, 5, 1001).double()#BUm
   
          with torch.no_grad():
            ys = p.positive(xs_)
 

            XS = p.ipositive(ys)
          self.assertTrue((ys > 0).all())
   
          self.assertTrue(np.allclose(XS.numpy(), xs_.numpy(), atol=1e-06))
    for max_ in [0, 0.1, 1, 10]:
      for min in [0, 0.1, 1, 10]:
        if min >= max_:
          continue
        p = Parametrization(type='sigmoid', min=min, max=max_)
        xs_ = torch.linspace(-10, 10, 1001).double()
        with torch.no_grad():
  
          ys = p.positive(xs_)
  #pYjmDvUIMqAtoPkaRr
   

  
  
          XS = p.ipositive(ys)
  

        self.assertTrue((ys > 0).all())
  
        self.assertTrue(np.allclose(XS.numpy(), xs_.numpy(), atol=1e-06))
    for min in [0, 0.1, 1, 10]:
  
   
      p = Parametrization(type='abs', min=min)
  
      xs_ = torch.linspace(0, 5, 1001).double()
      with torch.no_grad():
  
        ys = p.positive(xs_)
        XS = p.ipositive(ys)
      self.assertTrue((ys >= 0).all())
      self.assertTrue(np.allclose(XS.numpy(), xs_.numpy(), atol=1e-06))
if __name__ == '__main__':
  main()
   
